change location of types
