+++ packages/types/src/config.ts
@@ -2,7 +2,7 @@ import { z } from "zod";
 
 export const addNonProdDefault = <T>(
   schema: z.ZodSchema<T>,
-  defaultValue: T extends undefined ? never : T
+  defaultValue: T extends undefined ? never : T,
 ): z.ZodSchema<T> | z.ZodDefault<z.ZodType<T, z.ZodTypeDef, T>> => {
   if (NODE_ENV === NodeEnv.PROD) {
     return schema;
@@ -27,3 +27,34 @@ export enum ApiKey {
 
 const _apiSchema = z.enum([ApiKey.DEV, ApiKey.PROD_PUBLIC]);
 export const apiKeySchema = addNonProdDefault(_apiSchema, ApiKey.DEV);
+
+export const AztecNodeConfig = z.object({
+  name: z.string(),
+  url: z.string(),
+});
+
+export type AztecNodeConfig = z.infer<typeof AztecNodeConfig>;
+
+export const rpcNodePoolSchema = z
+  .string()
+  .default("local::http://localhost:8080")
+  .transform((str) => {
+    if (!str) {
+      return [];
+    }
+
+    const nodes: AztecNodeConfig[] = [];
+    const pairs = str.split(",");
+
+    for (const pair of pairs) {
+      const [key, url] = pair.split("::");
+      if (key && url) {
+        nodes.push({
+          name: key.trim(),
+          url: url.trim(),
+        });
+      }
+    }
+
+    return nodes;
+  });
